First Model D/D/1/K-1 semi-done!

+++ src/main/java/models/QueueSystemInput.java
@@ -6,11 +6,11 @@ import io.reactivex.annotations.Nullable;
 public class QueueSystemInput {
     private Double arrivalRate;
     private Double serviceRate;
-    private Double numberOfServers;
-    private Double systemCapacity;
+    private Integer numberOfServers;
+    private Integer systemCapacity;
 
     public QueueSystemInput(@NonNull Double arrivalRate, @NonNull Double serviceRate,
-                            @Nullable Double numberOfServers, @Nullable Double systemCapacity) {
+                            @Nullable Integer numberOfServers, @Nullable Integer systemCapacity) {
         this.arrivalRate = arrivalRate;
         this.serviceRate = serviceRate;
         this.numberOfServers = numberOfServers;
@@ -25,11 +25,11 @@ public class QueueSystemInput {
         return serviceRate;
     }
 
-    public Double getNumberOfServers() {
+    public Integer getNumberOfServers() {
         return numberOfServers;
     }
 
-    public Double getSystemCapacity() {
+    public Integer getSystemCapacity() {
         return systemCapacity;
     }
